Second Update: Fixed Functions
When running the first code an error may come up letting you know that the variables were not declared in the scop which is right. So the solution was to have the calculations happen in the main module/function where as there would be no need to pass variables. Code is now working smoothly. Feel free to run it! :)